refactor: refactor `theme.lua` and `util.lua`
